Updates to the configurator and admin->vehicles page (batch editing is still a WIP)

+++ webutil/src/main/java/com/inthinc/pro/configurator/model/ConfigurationSet.java
@@ -23,7 +23,7 @@ public class ConfigurationSet {
     	
     	for(Configuration c:configurations){
     		
-    		if(c.getConfigurationID() == configurationID) return c;
+    		if(c.getConfigurationID().intValue() == configurationID.intValue()) return c;
     	}
     	return null;
     }
